pyproject.toml
[project]
name = "cloud-cert-renewer"
version = "0.1.0"
description = "自动更新云服务HTTPS证书工具，支持CDN和负载均衡器（支持多云，当前支持阿里云）"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "alibabacloud-cdn20180510>=3.1.2",
    "alibabacloud-slb20140515>=3.1.2",
    "cryptography>=41.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-mock>=3.10.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "yamllint>=1.32.0",
    "pre-commit>=3.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["cloud_cert_renewer"]

[tool.black]
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'

[tool.ruff]
target-version = "py38"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["B018", "UP007"]  # 允许 Protocol 中使用 ...

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

